Function average_vector OK
The function in last cell (get_avg_vector(number_clusters, cluster_index)) returns the average vector corresponding to cluster n° cluster_index when kmeans was run with a total of number_clusters clusters.
It returns a vector v of length 134 (the number of aisles), such that v[3]=4.5 means "On average, users from this cluster bought 4.5 products in aisle of index 3+1=4". The "+1" comes from the fact that aisle ids start at 1. It also prints the number of users in the cluster, which may vary a lot from cluster to cluster (with 4 clusters, it varies from 150 000 to 4 000...)